Add Kasten again; Practise note taking
I renamed the note taking file to `Kasten` to highlight that all are
written in German now. There are multiple new links in the
Observatorium, the Europa Audiobooks are pretty cool. I also practiced
the decomposition process for note taking on McPherson book on Effective
Note Taking (of course ;)). Some general structure already lives in the
Kasten but is not yet complete.